# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "description": "Infrastructure inventory, observation, incident response, execution, and access control API.",
    "title": "KloudView API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "description": "Same-origin Server",
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/alert-rules": {
      "get": {
        "tags": [
          "alert-rules"
        ],
        "operationId": "getApiV1AlertRules",
        "summary": "GET /api/v1/alert-rules",
        "responses": {
          "2XX": {
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error response"
          }
        },
        "security": [
          {
            "headerSubject": []
          }
        ]
      },
      "post": {
        "tags": [
          "alert-rules"
        ],
        "operationId": "postApiV1AlertRules",
        "summary": "POST /api/v1/alert-rules",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "2XX": {
            "description": "Successful response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error response"
          }
        },
        "security": [
          {
            "headerSubject": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "properties": {
          "error": {
            "type": "object"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "agentBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "headerSubject": {
        "in": "header",
        "name": "X-KloudView-Subject",
        "type": "apiKey"
      }
    }
  }
}